# nixpkgs: poor mans doc blocks wip / rfc / etc