--- a/doc/lib.nix
+++ b/doc/lib.nix
@@ -103,14 +103,10 @@ let
           <refpurpose>${typeDocbook}</refpurpose>
         </refnamediv>
 
-        <refsect1>
-          <title></title>
-          <para>${mkGithubLink pos.file pos.line}</para>
-        </refsect1>
-
         <refsect1 role="description">
           <title>Description</title>
           <para>${description}</para>
+          <para>Source: ${mkGithubLink pos.file pos.line}</para>
         </refsect1>
 
         ${paramsDocbook}