--- a/lib/docs.nix
+++ b/lib/docs.nix
@@ -1,0 +1,12 @@
+{ lib }:
+{
+  mkDoc = x:
+
+    let
+      validate = {
+        description,
+        examples ? [],
+        type ? false
+      }: x;
+    in validate x;
+}