There has been a lot of chatter around doc blocks in Nix. Since Nix is almost purely expression based, it is likely very challenging to actually make that work. We might look in to how JSDoc handles it, as they have similar challenges I think.

However, that is complicated. I decided to get creative and try something simple out. Maybe you'll like it, maybe you'll hate it ;)

Basically, where we have:

```
  /* id returns its parameter
     Example:

    id x
    => x
*/
  id = x: x;
```

I changed it to:

```
  docs.id = mkDoc {
    description = "id returns its parameter";
    examples = [
      { title = "basic example";
        body = ''
          id x
          => x
        '';
    ];
  };
  id = x: x;
```

which get dumped to a docbook file and included in the manual. Currently very WIP, I spent an hour on this, looking for ways to solve this that are easier and more actionable than "we need a better parser."

Example result:
![screenshot-2017-7-15 nixpkgs contributors guide](https://user-images.githubusercontent.com/76716/28236083-7d0e1bc2-68ea-11e7-8777-6169b156be6f.png)

What do y'all think?

cc @joepie91 @domenkozar @edolstra @copumpkin @Profpatsch @zimbatm @vcunat @everyoneelse.