@bjornfor 2017-07-15
Tough one. It does make the generated documentation nice to read, but less so the source itself. On the other hand, it also makes the doc available in nix-repl, which is nice. I'm leaning towards +1 :-)