@zimbatm 2017-07-15
It would be nice if the function could embed the metadata somehow, that would allow to query any function and not have to worry of it's location or where it came from. But that would require a language change.

Even in it's current implementation I think that it's an improvement over writing XML.