@vcunat 2017-07-16
Theoretically, nix could be extended to parse documentation comments and expose them via some `builtins.getDoc`.  I don't expect that to be a very difficult change, but certainly not as easy as what you did.

Syntax within the docs seem a partially independent topic.  Maybe it would be simpler to read if we didn't solve syntax at nix layer (i.e. write just string lines) and left parsing to the docs generator, perhaps some standard format (markdown, asciidoc, ...).