@grahamc 2017-07-16
I quite like the idea of doing it without language features, so the functionality can evolve on its own. That said, I'm not particularly attached to any specific method.

I do actually like  https://github.com/NixOS/nixpkgs/pull/23505 better than this, if we could find a way to make it better... and of course, joepie thinks in-code docs are not good at all.