@Profpatsch 2017-07-17
I [started thinking about](https://github.com/jwiegley/hnix/issues/57#issuecomment-311553816) a comment-parsing implementation for the `hnix` parser of the nix language a few days back.

> I do actually like #23505 better than this, if we could find a way to make it more performant...

This indeed looks awesome. Not sure why `self` is passed into `__functor`, looks to me like it makes optimizations kind of hard.