@zimbatm 2017-07-19
> This indeed looks awesome. Not sure why self is passed into __functor, looks to me like it makes optimizations kind of hard.

nix doesn't bind the function to the object's context so there wouldn't be a way to parameterize the function. It's more obvious with `__toString` where a reference to `self` is necessary most of the time. It's a bit like python in that regard.