@zimbatm 2017-07-19
> I started thinking about a comment-parsing implementation for the hnix parser of the nix language a few days back.

If comments can be introspected they become part of the code and need to be tested like the rest of the code. It's probably better to keep comments free-form and add a notion of introspectable object annotation to the language. That way comments are reserved to the developer to better understand the code and it's possible to extend the language with more information. Nix values are already a big union struct so more data could be bolted-on.