@zimbatm 2017-07-19
For today, I think that @grahamc's proposal is a good solution for technical code that needs to stay in sync with the code. When/if the language gets extended the documentation can get refactored.