@edolstra 2017-07-19
Maybe an example like
```
  body = ''
    id x
    => x
  '';
```
should be
```
  input = "id 123";
  output = "123";
```
That way we get more flexibility in rendering the examples, and they can be tested automatically.