@ericsagnes 2017-07-23
For the record, I use a similar approach to #23505 in styx to generate library documentation and automatic tests, so far it works pretty well with no visible performance impact, but styx codebase size is nowhere close to nixpkgs.

It is still very basic and makes function declarations a lot more verbose, but having generated documentation and tests can really help.

Example of generated docs: https://styx-static.github.io/styx-site/documentation/library.html
Exemple of a test run (with a voluntary error in the expected result to show interesting output):

```
57 tests run.
- 56 success(es).
- 1 failure(s).

Failures details:

====================
lib.utils.sortBy, example number 1:

code:
---
sortBy "priority" "asc" [ { priority = 5; } { priority = 2; } ]
---

expected:
---
[ {
  priority = 2;
} {
  priority = 3;
} ]
---

got:
---
[ {
  priority = 2;
} {
  priority = 5;
} ]
---
====================
```