@grahamc 2017-09-17
Bringing this back from the dead, I realize I never commited docs.nix :( I'm rebasing on https://github.com/NixOS/nixpkgs/pull/27797, as I'm assuming that'll merge.