@grahamc 2017-09-17
Ok,You won't like _how_ I did this, but:

```
    examples =[
      (testEx {
        title = "constructing a concat function with foldr";
        test = rec {
          concat = fold (a: b: a + b) "z";
          ensure = concat [ "a" "b" "c" ] == "abcz";
        };
      })
```

produced:

![screenshot-2017-9-17 nixpkgs contributors guide](https://user-images.githubusercontent.com/76716/30517650-26ff8690-9b34-11e7-931d-79ee43741436.png)

and if the `ensure` is wrong, it'll fail to build the manual. ... just an experiment :)