@Profpatsch 2017-09-17
> and if the ensure is wrong, it'll fail to build the manual. ... just an experiment :)

Nice, I like that. I’ve been thinking about how/where we should integrate such tests that they fail as fast as possible. We should chat about this.