@grahamc 2017-09-20
@bjornfor @zimbatm @edolstra @Profpatsch @vcunat @ericsagnes please review the code, and the output:

 - http://gsc.io/nixos/nixpkgs/manual/#fn-lists-findSingle
 - http://gsc.io/nixos/nixpkgs/manual/#fn-attrsets-collect

Thus far I've received very positive feedback about the results. We can achieve these results in other ways. Is this a change we want to make? Do we want to document Fns like this?