@grahamc 2017-09-20
here are some totally bogus benchmarks to see if I slowed things down a lot:

master:
```
nix-instantiate ./nixos/tests/*.nix  23.40s user 0.59s system 132% cpu 18.070 total
nix-instantiate ./nixos/tests/*.nix  24.51s user 0.61s system 146% cpu 17.193 total
nix-instantiate ./nixos/tests/*.nix  24.44s user 0.57s system 147% cpu 17.006 total
```

this branch:
```
nix-instantiate ./nixos/tests/*.nix  25.18s user 0.83s system 123% cpu 21.044 total
nix-instantiate ./nixos/tests/*.nix  25.37s user 0.54s system 153% cpu 16.854 total
nix-instantiate ./nixos/tests/*.nix  24.65s user 0.54s system 150% cpu 16.760 total
nix-instantiate ./nixos/tests/*.nix  23.96s user 0.60s system 150% cpu 16.337 total
```