@zimbatm 2017-09-20
I think that the latest changes go a bit too far in terms of verbosity, the function description now takes a full page in my browser. In most cases it's enough to document the parameter names and types and then expand in the description and examples in case it's not clear what those parameters are doing.

How about changing the output to be `lists.findSingle — pred: (a -> Bool) -> default: a -> multiple: a -> [a] -> a`? With named arguments it makes it easier to reference them in the function body.