@zimbatm 2017-09-20
I don't know if there is a type notation that could be closer to actually nix code
```
lists.findSingle =
  pred: # (a -> Bool)
  default: # a
  multiple: #a
  ; #=> a
```